chore(cmd): Move FlagClientAgentPort and flagTargetDefaultPort to uint16

+++ internal/cmd/commands/clientagentcmd/sessions.go
@@ -58,7 +58,7 @@ func (c *SessionsCommand) Flags() *base.FlagSets {
 		Usage:  "Instead of executing the request, print an equivalent cURL command string and exit.",
 	})
 
-	f.Uint64Var(&base.Uint64Var{
+	f.Uint16Var(&base.Uint16Var{
 		Name:    "client-agent-port",
 		Target:  &c.FlagClientAgentPort,
 		Default: 9300,
