+++ internal/cmd/commands/clientagentcmd/status.go
@@ -54,7 +54,7 @@ func (c *StatusCommand) Flags() *base.FlagSets {
 		Usage:  "Instead of executing the request, print an equivalent cURL command string and exit.",
 	})
 
-	f.Uint64Var(&base.Uint64Var{
+	f.Uint16Var(&base.Uint16Var{
 		Name:    "client-agent-port",
 		Target:  &c.FlagClientAgentPort,
 		Default: 9300,
